requirements.txt
aioredis==1.1.0
aniso8601==3.0.2
asgiref==2.3.2
astroid==2.0.4
async-timeout==2.0.1
attrs==18.1.0
autobahn==18.8.1
Automat==0.7.0
autopep8==1.3.5
channels==2.1.0
channels-redis==2.3.0
constantly==15.1.0
daphne==2.2.2
dj-database-url==0.5.0
Django==2.0.7
django-cors-headers==2.4.0
django-environ==0.4.5
django-graphql-jwt==0.1.10
django-heroku==0.3.1
graphene==2.1.2
graphene-django==2.0.0
graphql-core==2.0
graphql-relay==0.4.5
gunicorn==19.9.0
hiredis==0.2.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
iso8601==0.1.12
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
msgpack==0.5.6
promise==2.1
psycopg2==2.7.5
pycodestyle==2.4.0
PyHamcrest==1.9.0
PyJWT==1.6.4
pylint==2.1.1
pytz==2018.5
rope==0.11.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
Twisted==18.7.0
txaio==18.7.1
typed-ast==1.1.0
typing==3.6.4
whitenoise==3.3.1
wrapt==1.10.11
yapf==0.23.0
zope.interface==4.5.0
